Count a channel's event types in 16 bits everywhere, not just on the wire
The 0x80 count and the 0x85 index went to two bytes, but the code that
produces the index did not: _findEventType walked the pool with a byte
ordinal, so the 256th type on a channel would have answered 0 and every
later one would have repeated an earlier index. Two bytes on the wire
carrying a number that wrapped before it got there.

_flashCsvOptionCount had the same shape - a byte counter over a CSV, so
addEventChannel(F("a,b,c,...")) with more than 255 fields would have
declared 44 of them and said nothing. Widened along with the field loop
that walks it.

Neither is reachable by accident: both need a single flash literal with
256 names in it. But both were silent, and the wire is 16-bit now, so the
counters that feed it should be too.

Select options share that counter and are unaffected in practice -
getSelectOptionNameAt() still takes a byte index, which is its own limit
and a sensible one for a control a person picks from.

+++ src/BlaeckSerial.cpp
@@ -1408,12 +1408,12 @@ BlaeckTextCommandRef BlaeckSerial::onTextCommand(const char *command, BlaeckComm
   return BlaeckTextCommandRef(this, (int16_t)_registerCommand(command, handler, BLAECK_CMD_TEXT));
 }
 
-byte BlaeckSerial::_flashCsvOptionCount(const __FlashStringHelper *csv)
+uint16_t BlaeckSerial::_flashCsvOptionCount(const __FlashStringHelper *csv)
 {
   if (csv == nullptr)
     return 0;
   PGM_P p = reinterpret_cast<PGM_P>(csv);
-  byte count = 1;
+  uint16_t count = 1;
   bool any = false;
   byte c;
   while ((c = pgm_read_byte(p++)) != 0)
@@ -2773,8 +2773,8 @@ void BlaeckSerial::_addEventTypesCsv(uint16_t channelIndex, const __FlashStringH
   // One pool entry per field, all pointing at the same flash string. Appended in
   // order, so a field's position is its wire index - the same rule call order gives
   // addEventType().
-  byte fieldCount = _flashCsvOptionCount(eventTypes);
-  for (byte f = 0; f < fieldCount; f++)
+  uint16_t fieldCount = _flashCsvOptionCount(eventTypes);
+  for (uint16_t f = 0; f < fieldCount; f++)
   {
     if (!_ensureEventTypeTable() || _eventTypeCount >= _eventTypeSlots())
     {
@@ -2949,7 +2949,7 @@ int BlaeckSerial::_findEventType(uint16_t channelIndex, const __FlashStringHelpe
   // F() literals at different addresses. Both operands live in flash, so
   // neither strcmp() nor strcmp_P() applies (the latter reads its first
   // argument from RAM) — read both sides with pgm_read_byte().
-  byte index = 0;
+  uint16_t index = 0;
   for (uint16_t i = 0; i < _eventTypeCount; i++)
   {
     if (_eventTypes[i].channelIndex != channelIndex)
@@ -3317,7 +3317,7 @@ byte BlaeckSerial::_validateTypedCommand(uint16_t handlerIndex)
   }
   else if (e.kind == BLAECK_CMD_SELECT)
   {
-    byte count = _flashCsvOptionCount(e.options);
+    uint16_t count = _flashCsvOptionCount(e.options);
 
     // Accept either an option name (case-insensitive) or a numeric index.
     long idx = _flashCsvIndexOf(e.options, v);
